[#210] Added factService, left only ml service specific methods in mlService
